Created separate folder in graphics/ for Easy3D-specific classes
