+++ src/graphics/easy3d/Easy3DGraphicsScene.cpp
@@ -1,10 +1,10 @@
-#include "graphics/Easy3DGraphicsScene.hpp"
-#include "graphics/Easy3DMeshGraphicsObject.hpp"
-#include "graphics/Easy3DSphereGraphicsObject.hpp"
-#include "graphics/Easy3DBoxGraphicsObject.hpp"
-#include "graphics/Easy3DCylinderGraphicsObject.hpp"
-#include "graphics/Easy3DVirtuosoArmGraphicsObject.hpp"
-#include "graphics/Easy3DVirtuosoRobotGraphicsObject.hpp"
+#include "graphics/easy3d/Easy3DGraphicsScene.hpp"
+#include "graphics/easy3d/Easy3DMeshGraphicsObject.hpp"
+#include "graphics/easy3d/Easy3DSphereGraphicsObject.hpp"
+#include "graphics/easy3d/Easy3DBoxGraphicsObject.hpp"
+#include "graphics/easy3d/Easy3DCylinderGraphicsObject.hpp"
+#include "graphics/easy3d/Easy3DVirtuosoArmGraphicsObject.hpp"
+#include "graphics/easy3d/Easy3DVirtuosoRobotGraphicsObject.hpp"
 #include "config/simobject/MeshObjectConfig.hpp"
 
 #include "simobject/MeshObject.hpp"
